add: metadados em paginas de Doar, Home, Noticias, Projetos, Voluntario, Sobre e Transparência
