+++ app/components/faq/DonationFAQ.tsx
@@ -5,8 +5,7 @@ import {
   AccordionSummary,
   AccordionDetails,
   Typography,
-  Box,
-  Link,
+  Box
 } from '@mui/material'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import { siteConfig } from '@/app/data/site.config'
